moved setUp and tearDown methods to CoreTestCase.java
