+++ src/lib/CoreTestCase.java
@@ -1,0 +1,46 @@
+package lib;
+
+import io.appium.java_client.AppiumDriver;
+import io.appium.java_client.android.AndroidDriver;
+import junit.framework.TestCase;
+import org.openqa.selenium.ScreenOrientation;
+import org.openqa.selenium.WebElement;
+import org.openqa.selenium.remote.DesiredCapabilities;
+
+import java.net.URL;
+
+public class CoreTestCase extends TestCase {
+
+    protected AppiumDriver driver;
+    private static String AppiumURL = "http://127.0.0.1:4723/wd/hub";
+
+    @Override
+    public void setUp() throws Exception {
+
+        super.setUp();
+
+        DesiredCapabilities capabilities = new DesiredCapabilities();
+
+        capabilities.setCapability("platformName","Android");
+        capabilities.setCapability("deviceName","AndroidTestDevice");
+        capabilities.setCapability("platformVersion","8.0");
+        capabilities.setCapability("automationName","Appium");
+        capabilities.setCapability("appPackage","org.wikipedia");
+        capabilities.setCapability("appActivity",".main.MainActivity");
+        capabilities.setCapability("app","D:\\JavaAppiumAutomation\\apks\\org.wikipedia.apk");
+
+        driver = new AndroidDriver<>(new URL(AppiumURL), capabilities);
+        driver.rotate(ScreenOrientation.PORTRAIT);
+
+        WebElement skip_button = driver.findElementById("fragment_onboarding_skip_button");
+        skip_button.click();
+    }
+
+    @Override
+    public void tearDown() throws Exception {
+        driver.quit();
+
+        super.tearDown();
+    }
+
+}
